Adds cart middleware and POST for /charges. *needs refactor*
Adds Cart middleware which stores the complete cart to res.locals
for later access.
Adds a POST method for the /charges route. This needs a refactor -
currently it calls the mongoose model directly. Move that logic
into mongoose-queries.js
Also, mongoose-queries and sequelize-data into the same folder and
standardize the file names.